Bugfix in Loader; Added Help
There was a bug in the `loadArchive` recursive method - when loading arrays, it pushed the newly loaded `DndNode` objects into the `value` field of the parent, but the constructor of `DndNode` already does this. Hence, there was a repetition which caused the number of child nodes of array members to double.